+++ formio/models/formio_builder.py
@@ -128,6 +128,12 @@ class Builder(models.Model):
         - Relative URL is also supported e.g. /web/login
         """
     )
+    portal_scroll_into_view_selector = fields.Char(
+        string='Portal Scroll Into View Selector',
+        copy=False,
+        tracking=True,
+        help="Especially for long wizard pages upon prev/next page. This scrolls an element (CSS selector) into the visible area of the browser window."
+    )
     public = fields.Boolean("Public", tracking=True, help="Form is public accessible (e.g. used in Shop checkout, Events registration")
     public_url = fields.Char(string='Public URL', compute='_compute_public_url')
     public_save_draft_done_url = fields.Char(
@@ -153,6 +159,12 @@ class Builder(models.Model):
         tracking=True)
     public_access_interval_number = fields.Integer(default=30, tracking=True, help="Public access to submitted Form shall be rejected after expiration of the configured time interval.")
     public_access_interval_type = fields.Selection(list(_interval_selection.items()), default='minutes', tracking=True)
+    public_scroll_into_view_selector = fields.Char(
+        string='Public Scroll Into View Selector',
+        copy=False,
+        tracking=True,
+        help="Especially for long wizard pages upon prev/next page. This scrolls an element (CSS selector) into the visible area of the browser window."
+    )
     view_as_html = fields.Boolean("View as HTML", tracking=True, help="View submission as a HTML view instead of disabled webform.")
     show_form_title = fields.Boolean("Show Form Title", tracking=True, help="Show Form Title in the Form header.", default=True)
     show_form_id = fields.Boolean("Show Form ID", tracking=True, help="Show Form ID in the Form header.", default=True)
@@ -612,6 +624,7 @@ class Builder(models.Model):
             'cdn_base_url': self._cdn_base_url(),
             'portal_save_draft_done_url': self.portal_save_draft_done_url,
             'portal_submit_done_url': self.portal_submit_done_url,
+            'portal_scroll_into_view_selector': self.portal_scroll_into_view_selector,
             'wizard_on_change_page_save_draft': self.wizard and self.wizard_on_change_page_save_draft,
             'submission_url_add_query_params_from': self.submission_url_add_query_params_from,
         }
@@ -623,6 +636,7 @@ class Builder(models.Model):
             'cdn_base_url': self._cdn_base_url(),
             'public_save_draft_done_url': self.public_save_draft_done_url,
             'public_submit_done_url': self.public_submit_done_url,
+            'public_scroll_into_view_selector': self.public_scroll_into_view_selector,
             'wizard_on_change_page_save_draft': self.wizard and self.wizard_on_change_page_save_draft,
             'submission_url_add_query_params_from': self.submission_url_add_query_params_from,
         }