+++ formio/static/src/js/form/formio_form.js
@@ -57,6 +57,10 @@ export class OdooFormioForm extends Component {
         // Implemented in specific (*_app.js) classes.
     }
 
+    scrollIntoView() {
+        // Optionally implemented in specific (*_app.js) classes.
+    }
+
     getDataUrl(compObj) {
         // Possibly overridden in specific (*_app.js) classes.
         return '/formio/form/', self.formUuid, compObj.data.url;
@@ -106,6 +110,10 @@ export class OdooFormioForm extends Component {
                         this.submitUrl = this.wizardSubmitUrl + this.formUuid + '/submit';
                     }
                     self.hideOverlay();
+                    if (window.self !== window.top) {
+                        // window.parent.postMessage({odooFormioMessage: 'formioScroll', params: self.params});
+                        self.scrollIntoView();
+                    }
                 });
             });
         }